Corrections to wallet generation
The generator for 'Inductive' was generating invalid wallets. I've tightened
the tests to check for this; in particular to check that the 'pending'
transactions we submit are valid (use only inputs that are available to the
wallet). In this commit I make some corrections, and the tests now _sometimes_
pass, but not yet always. In this commit:

* Extend some of the debugging infrastructure; allow for additional error
  messages in 'Validated' and for comments in 'Transaction' in the DSL, so that
  we can insert some information about where tranasctions are coming from etc.
* Introduce a `walletBoot` function that "boots up" a wallet; this really
  just means initializing the wallet and then applying the genesis block.
  This is important because otherwise the wallet lacks context to interpret
  the rest of the chain.
* This is then accompanied by a change to 'Inductive': renamed
  'WalletEmpty' to 'WalletBoot', and added a transaction argument
  (the bootstrap transaction). Propagated this change throughout.
* The 'ours' that was created in the test was a placeholder only.
  Changed 'genFromBlockchainPickingAccounts' to return the addresses
  it picked, and generate 'ours' from that.

All of the above were fixes to the infrastructure and to the tests themselves; the following are fixes to the wallet generation proper:

* New pending transactions were inserted using `Map.insertWith (<>)`, but
  that inserts them _before_ the block, rather than after. Changed to
  `Map.insertWith (flip (<>))`.
* After the previous fix, pending transactions were inserted "between" the
  blocks in which their (last) inputs became available and the block in which
  they were confirmed; but this would break for transactions in which those two
  blocks happen to coincide (i.e., transactions that are confirmed in the same
  block as their inputs). This would lead to transactions being inserted in
  `(lo, hi - 1)`, for `lo == hi`, i.e., in `lo - 1`, which is not correct.  We
  now skip such transactions (which is the only thing we _can_ do; such
  transactions could never be pending from the point of view of this wallet).
* Also in the previous fix the generation of pending transactions was
  fixed to look for transactions _from_ addresses owned by the wallet rather
  than transactions _to_ these addresses, but it was only checking whether
  _some_ of the addresses of the transactions were owned by the wallet rather
  than _all_.

The 'Inductive' generation is still broken even with these fixes however. Since
there is no connection between the pending transactions inserted in the first
pass, and the "never to be confirmed" transactions in the second pass (in
'synthesizeTransactions'), the latter pass may be using inputs that are
unavailable because they are used by pending transactions inserted previously.
I will try to fix this next, as well as try to avoid the `O(N^2)` computations
this function currently does to recompute the ledger's UTxO over and over
again.